Exercise 7.4 done
