+++ Part-07/routed-anecdotes/src/components/Footer.jsx
@@ -1,0 +1,12 @@
+const Footer = () => (
+  <div>
+    Anecdote app for <a href="https://fullstackopen.com/">Full Stack Open</a>.
+    See{" "}
+    <a href="https://github.com/fullstack-hy2020/routed-anecdotes/blob/master/src/App.js">
+      https://github.com/fullstack-hy2020/routed-anecdotes/blob/master/src/App.js
+    </a>{" "}
+    for the source code.
+  </div>
+);
+
+export default Footer;
