+++ Part-07/routed-anecdotes/src/components/Menu.jsx
@@ -1,0 +1,22 @@
+import { Link } from "react-router-dom";
+
+const Menu = () => {
+  const padding = {
+    paddingRight: 5,
+  };
+  return (
+    <div>
+      <Link style={padding} to="/">
+        anecdotes
+      </Link>
+      <Link to="/create" style={padding}>
+        create new
+      </Link>
+      <Link to="/about" style={padding}>
+        about
+      </Link>
+    </div>
+  );
+};
+
+export default Menu;
