+++ Part-07/routed-anecdotes/src/components/Notification.jsx
@@ -1,0 +1,14 @@
+const Notification = ({ notification }) => {
+  const style = {
+    border: "solid",
+    padding: 10,
+    borderWidth: 1,
+    marginBottom: 5,
+  };
+  if (notification === null) {
+    return null;
+  }
+  return <div style={style}>{notification}</div>;
+};
+
+export default Notification;
